move chat parsing to frontend
