+++ app/lib/parse-stream-message.ts
